Move Terraform class to lib/cp_env/terraform.rb
This is a step towards imposing a more idiomatic
ruby-project file structure.
