+++ lib/cp_env.rb
@@ -1,0 +1,3 @@
+require "open3"
+
+require File.join(File.dirname(__FILE__), "cp_env", "terraform")
